Hyperlinked "Smart Search plugin not enabled"
This PR changes the language file for the **Smart Searchr** (com_finder) so that the "Smart Search content plugin is not enabled. Changes to content will not update the Smart Search index if you do not enable this plugin." contains an hyperlink to the Plugin Manager.

## Test instructions
In back-end > Components > **Smart Search**
If the Smart Search - Content Plugin is not enabled, it will display a message "Smart Search content plugin is not enabled. Changes to content will not update the Smart Search index if you do not enable this plugin."



### This PR changes the language string 
so that the warning message has a hyperlink to the Plugin Manager,
including a filter option for the name "Content - Smart Search".


### Result of that hyperlink
If you click on the hyperlink you will be redirected to the Plugin Manager,
which filters the "Content - Smart Search" so you can easily enable it.